Fixed the Sizing for ranger prefabs, Added Killzones, add debug tools for first build
I fixed danielles test scene to be to the original scale of the ranger
prefabs. I also played with the jumping stats of the rangers to work
with the stage. Killzones are now a thing so if the ranger falls off of
the stage they will die. I have also add some debug tools for the first
playtest. mainly the ability to reset and to exit the game. I have also
implemented a basic dodge.