Finished Blue ranger Ice Blast Attack
I finished coding in the blue ranger's super "The Ice Blast" attack. I
will only be able to be used if the blue ranger has enough super. The
attack will freeze enemy for a certain amount of time and it will apply
freeze damage to player while it is frozen.  I have also added code into
the melee and ice shuriken attack to add super value to the rangers
super meter when the attacks are successful.