changed the way attacking works and updated prefabs
I updated the way attacking works and simplified it. instead of having
an attack range, there is now a hitrange child of players. so now the
range for each attack can be easily edited by modifying the collider for
that. the attack method now takes in the enemy player in range from the
hitbox check, then checks if the corresponding input is being pressed.