Game UI is now Implemented
the Game UI is now implemented in the game. It will track the players
health, assign the correct profile based on players character, and track
the super status. super bar will change color when super is ready. If a
player dies it will replace their profile image with dead ranger image.