Setup Character Selection Scene
I have started to setup what the character selection screen will look
like. currently each player will have a ranger on a platform and under
that platform will be the character selector tool. you will change
characters by using the right and left bumper. the screen is split up
into 4 quadrants one for each player. The Players will have to hit the a
button to join and once two players have joined they will see the option
to start the game. Currently the scene is setup with all the assets but
the interaction and code for it has not been done yet.