Fixed Blue Ranger frozen tint
I fixed the blue rangers frozen tint effect. the problem was that we
changed from one sprite render for the whole body to multiple sprite
renders for the body. So I changed the code to now get all sprite
renders in the ranger and change all the tints to frozen and back to
white when unfrozen.