Made font consistent + fixed Sprite ordering + unconfirm instruct
I went through the entire game and made all of the fonts to be
consistent with the action comic font we are using. I went into the
player selection and changed the player ready scroll image to have
instructions on how to unconfirm your character. In the loading scene I
swapped out the loading text, remove the position keyframe in the
shuriken load animation, and I fixed a bug with the continue btn not
being highlighted. In the stage 1 scene I change the sprite ordering
layer so that everything overlaps correctly, same deal with the win
scene. I also made it so the win scene can go back to the main menu
instead of exiting the game.