Fixed poison cloud naming bug and Green ranger vine bug
I fixed both of the game breaking bugs. The poison cloud was not working
because the script was misnamed on my end which was causing it not to
compile. As for the green ranger vine attack the OnTriggerEnter2D was
being called before the Start() which was causing null refrences. both
have been fixed.